Split remote volume commands into module
